requirements.txt
requests
tqdm
urlpath
GitPython
rich
colorama
pydantic